Separate frontend from backend
Assembly and machine code parsing code was extracted from assembler
and linker classes.
