+++ mvn-assembler/Cargo.toml
@@ -8,10 +8,9 @@ edition = "2021"
 [dependencies]
 indoc = "1.0.6"
 nom = "7.1.1"
-nom_locate = "4.0.0"
-dotenv_codegen = "0.15.0"
-utils = { path = "../mvn-utils", package = "mvn-utils" }
 annotate-snippets = { version = "0.9.1", features = ["color"] }
+assembly = { path = "../mvn-assembly", package = "mvn-assembly" }
+utils = { path = "../mvn-utils", package = "mvn-utils" }
 
 [dev-dependencies]
 pretty_assertions = "1.2.1"
