+++ mvn-assembler/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod parser;
 pub mod processor;
-pub mod types;
 pub mod writer;
+
+pub use assembly::*;
