+++ mvn-assembly/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "mvn-linker"
+name = "mvn-assembly"
 version = "0.1.0-alpha"
 edition = "2021"
 
@@ -8,9 +8,10 @@ edition = "2021"
 [dependencies]
 indoc = "1.0.6"
 nom = "7.1.1"
+nom_locate = "4.0.0"
+dotenv_codegen = "0.15.0"
 utils = { path = "../mvn-utils", package = "mvn-utils" }
-machine-code = { path = "../mvn-machine-code", package = "mvn-machine-code" }
-assembly = { path = "../mvn-assembly", package = "mvn-assembly" }
+annotate-snippets = { version = "0.9.1", features = ["color"] }
 
 [dev-dependencies]
 pretty_assertions = "1.2.1"
