+++ mvn-assembly/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod parser;
+pub mod types;
