+++ mvn-assembly/src/parser/label.rs
