+++ mvn-assembly/src/types/label.rs
