+++ mvn-assembly/src/types/mneumonic/normal.rs
